move ner pos in token classification module
